fixing errors
I spent the majority of the day debugging. the key issues with my code was that there was an incorrect order of functions in a redux file, therefore coulld not initalize.

then the next big issue was the incorrect attempt at manipulating JSON data. it is best if the data is manipulated within the component for printing out multiple posts.

the useSelector() only uses a selectFeed for example, this method is used to access the state of that slice, thus over specfifcation will lead to an error